add upgrade step fixing existing quota usages.
